% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_pre_post_processing.R
\name{clean_latin_text}
\alias{clean_latin_text}
\title{Remove Non-Latin Characters from Text}
\usage{
clean_latin_text(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector.
}
\description{
Remove Non-Latin Characters from Text
}
