% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_targeted_sentiment.R
\name{get_targeted_sentiment}
\alias{get_targeted_sentiment}
\title{Get Targeted Sentiment
This function performs targeted sentiment analysis on a given dataset using a specified pre-trained model.}
\usage{
get_targeted_sentiment(
  data,
  text_col = "translation",
  entity_col = "target",
  id_col = "id",
  conda_env_name = "r-sentiner",
  verbose = TRUE,
  model = "MoritzLaurer/deberta-v3-base-zeroshot-v2.0",
  batch_size = 32L,
  entailment_index = 0L,
  contradiction_index = 1L,
  fp16 = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame or data table containing the text and target entities for sentiment analysis.}

\item{text_col}{The name of the column in \code{data} that contains the text
to be analyzed. Default is "translation".}

\item{entity_col}{The name of the column in \code{data} that contains the target
entities for sentiment analysis. Default is "target".}

\item{id_col}{The name of the column in \code{data} that contains unique identifiers
for each row. Default is "id".}

\item{conda_env_name}{The name of the conda environment where the sentiment analysis
model is installed. Default is "r-sentiner".}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{model}{The pre-trained model to use for sentiment analysis. Default is
"MoritzLaurer/deberta-v3-base-zeroshot-v2.0".}

\item{batch_size}{The batch size to use during inference. Default is 32L.}

\item{entailment_index}{The index of the entailment class in the model's output. Default is 0L.}

\item{contradiction_index}{The index of the contradiction class in the model's output. Default is 1L.}

\item{fp16}{Logical indicating whether to use half-precision floating point
during inference. Default is TRUE.}

\item{...}{Additional arguments to pass to the \code{initialize_sentiner} function.}
}
\value{
A data table containing the original data along with the predicted
sentiment, sentiment confidence, sentiment model, and sentiment datetime.
}
\description{
Get Targeted Sentiment
This function performs targeted sentiment analysis on a given dataset using a specified pre-trained model.
}
