% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_installs.R
\name{install_sentiner}
\alias{install_sentiner}
\title{Install GLiNER in a conda Environment}
\usage{
install_sentiner(
  python_version = "3.11",
  conda_env_name = "r-sentiner",
  check_path = TRUE,
  ask = TRUE,
  force = FALSE,
  verbose = TRUE,
  conda_path = NULL
)
}
\arguments{
\item{python_version}{Character string. Python version to use when creating
the conda environment. Default is \code{"3.11"}.}

\item{conda_env_name}{Character string. Name of the conda environment to
create or activate. Default is \code{"r-sentiner"}.}

\item{check_path}{Logical. Checks whether the installed conda is on PATH. Default is TRUE.

The function installs the following dependencies if FastText is not
available: \code{onnxruntime}, \code{numpy}, \code{protobuf}, \code{sentencepiece},
\code{torch}, \code{tqdm}, \code{transformers}, \code{cython},
\code{pysimdjson}, \code{accelerate}, and \code{gliner}.}

\item{ask}{Logical. If \code{TRUE}, ask before installing Miniconda. Default
is \code{TRUE}.}

\item{force}{Logical. If \code{TRUE}, force reinstallation of the conda
environment and Python packages. Default is \code{FALSE}.}

\item{verbose}{Logical. If \code{TRUE}, print progress messages during
installation. Default is \code{TRUE}.}

\item{conda_path}{Optional character string. Path to a conda installation. If
\code{NULL}, the default conda installation is used.}
}
\description{
Installs the GLiNER library and its dependencies in a conda environment
managed through \code{reticulate}. It creates or activates the specified
environment, installs PyTorch (with CUDA if available, otherwise CPU),
installs FastText (using a precompiled wheel on Windows), and finally
installs gliner with all required dependencies.

Note: Calling GliNER from reticulate is hacky, as there is a persistent
DLL Error. At the moment, I implemeted a workaround that calls gliner
externally, without embedding it via reticulate.
}
