% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_pre_post_processing.R
\name{is_latin_script_lang}
\alias{is_latin_script_lang}
\title{Check Whether a Language Uses Latin Script}
\usage{
is_latin_script_lang(lang)
}
\arguments{
\item{lang}{Character vector of ISO 639-1 language codes.}
}
\value{
Logical vector.
}
\description{
Check Whether a Language Uses Latin Script
}
