% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_preprocessing.R
\name{replace_emoji_with_name}
\alias{replace_emoji_with_name}
\title{Replace Emojis with Names}
\usage{
replace_emoji_with_name(text_vec)
}
\arguments{
\item{text_vec}{Character vector of texts.}
}
\value{
Character vector with emojis replaced by names.
}
\description{
Replaces all emojis in a character vector with their textual names (in :name: style).
}
